requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
curl_cffi==0.6.2
pyrebase4==4.7.1